Finished lecture 4 exercise problems.
The second exercise problem had many sub questions, and I ended up
creating a few files specifically tied to that question. So, to organize
things better, I put all the files related to that problem in their own
subdirectory called q2.

The last exercise problem was a little tough, and I did not answer the
final portion exactly. Maybe come back to this for a future attempt.