Added lecture 6 exercises.
A word of warning when it comes to attempting to rewrite history. When
first working on the exercises and testing out my code related to the
changing commit history exercise, I accidentally had the branch I was
working on lose its parents, and all files beside the Lecture6 directory
were removed.

Workarounds were made, but as stated in the lecture 6 markdown file,
caution is advised when performing actions like this.