Finished lecture 7 exercises.
Will need to come back at a later point to learn more about `cgroups`,
as I feel like the documentation did not do a well enough job explaining
how to use the v2 compared to v1. Also worth noting is that there are
other tools for using and creating `cgroups`, so that might make matters
easier.

Also consider learning about wireshark in the future, but for right now,
eh.